Turn off verbose output
The debug output on stdout was getting way too chatty for my tastes, and it seemed like a performance liability as well. Turning it off, but in a way that makes it easy to turn on later with some kind of config option when we decide how to implement that.